api/python/requirements.txt
anyio==3.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
distro==1.8.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Session==0.4.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
modern-treasury==1.11.0
pydantic==1.10.7
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==2.0.3
Werkzeug==2.2.3